+++ cstar/execution/handler.py
@@ -1,3 +1,4 @@
+import time
 from abc import ABC, abstractmethod
 from enum import Enum, auto
 
@@ -42,11 +43,51 @@ class ExecutionStatus(Enum):
 
 
 class ExecutionHandler(ABC):
-    @abstractmethod
-    def updates(self):
-        pass
-
     @property
     @abstractmethod
     def status(self):
         pass
+
+    def updates(self, seconds=10):
+        """Provides updates from the job's output file as a live stream for `seconds`
+        seconds (default 10).
+
+        If `seconds` is 0, updates are provided indefinitely until the user interrupts the stream.
+        """
+
+        if self.status != ExecutionStatus.RUNNING:
+            print(
+                f"This job is currently not running ({self.status}). Live updates cannot be provided."
+            )
+            if (self.status in {ExecutionStatus.FAILED, ExecutionStatus.COMPLETED}) or (
+                self.status == ExecutionStatus.CANCELLED and self.output_file.exists()
+            ):
+                print(f"See {self.output_file.resolve()} for job output")
+            return
+
+        if seconds == 0:
+            # Confirm indefinite tailing
+            confirmation = (
+                input(
+                    "This will provide indefinite updates to your job. You can stop it anytime using Ctrl+C. "
+                    "Do you want to continue? (y/n): "
+                )
+                .strip()
+                .lower()
+            )
+            if confirmation not in {"y", "yes"}:
+                return
+
+        try:
+            with open(self.output_file, "r") as f:
+                f.seek(0, 2)  # Move to the end of the file
+                start_time = time.time()
+
+                while seconds == 0 or (time.time() - start_time < seconds):
+                    line = f.readline()
+                    if line:
+                        print(line, end="")
+                    else:
+                        time.sleep(0.1)  # 100ms delay between updates
+        except KeyboardInterrupt:
+            print("\nLive status updates stopped by user.")
