+++ examples/cstar_example.py
@@ -10,7 +10,7 @@ roms_marbl_case = cstar.Case.from_blueprint(
     blueprint="cstar_blueprint_roms_marbl_example.yaml",
     caseroot="roms_marbl_example_case/",
     start_date="20120103 12:00:00",
-    end_date="20120104 12:00:00",
+    end_date="20120103 18:00:00",
 )
 
 ## In a python session, execute:
